requirements.txt
streamlit
numpy
matplotlib
pandas
pillow 
astor==0.8.1
plotly
requests
ecoinvent_interface
xlsxwriter